Update SQL-BackupScript.PS1
This was a non-functional script so I got it working and improved it. Some of the logic had to get corrected or was repetitive. Added some logic to create the log directory if it doesn't exist, and set the default path for the log directory to be a subdirectory of the backup directory. enforced the backup retention parameter to be a positive value so no longer needed an if condition to evaluate this. Minor spelling corrections. Converted script to be a parameterized script which can still be called from a scheduled task. Removed some comments that explained obvious things that were happening.